+++ branches/jAER_2.0/src/main/java/net/sf/jaer2/eventio/events/XYPositionEvent.java
@@ -3,8 +3,8 @@ package net.sf.jaer2.eventio.events;
 public class XYPositionEvent extends Event {
 	private static final long serialVersionUID = 5576838970200124104L;
 
-	public int x;
-	public int y;
+	public short x;
+	public short y;
 
 	public XYPositionEvent(final int ts) {
 		super(ts);
